Working version.
