+++ sort/gnome_sort/cpp/main.cpp
@@ -5,10 +5,10 @@
 int main()
 {
     std::vector<double> arr = {0.0, 0.2, 0.01, 0.6, 0.0, 0.15, 0.01, 10, 0.54, 0.2};
-    unsigned int n = arr.size();
+    size_t n = arr.size();
 
     gnome_sort(arr, n);
-
+    
     for (double val : arr)
     {
         std::cout << val << ' ';
